refactor: adiciona tratamento de erros e ajustes gerais no código
