+++ src/assets/helpers/pets.ts
@@ -1,0 +1,35 @@
+import fish from '@/assets/img/pets/1_fish.svg'
+import frog from '@/assets/img/pets/3_frog.svg'
+import snake from '@/assets/img/pets/4_snake.svg'
+import turtle from '@/assets/img/pets/2_turtle.svg'
+
+export const aquariumPets = [
+  {
+    id: 1,
+    name: 'Peixe',
+    img: fish,
+    selected: false,
+    quantity: 0,
+  },
+  {
+    id: 2,
+    name: 'Tartaruga',
+    img: turtle,
+    selected: false,
+    quantity: 0,
+  },
+  {
+    id: 3,
+    name: 'Sapo',
+    img: frog,
+    selected: false,
+    quantity: 0,
+  },
+  {
+    id: 4,
+    name: 'Cobra',
+    img: snake,
+    selected: false,
+    quantity: 0,
+  },
+]
